// lib/session.ts
//session.ts file is used to store all the data about the current logged in user
import { getServerSession } from "next-auth/next";
import { NextAuthOptions, User } from "next-auth";
import { AdapterUser } from "next-auth/adapters";
import GoogleProvider from "next-auth/providers/google";
import jsonwebtoken from "jsonwebtoken";
import { JWT } from "next-auth/jwt";
import { SessionInterface, UserProfile } from "@/ccmmon.types";
import { createUser, getUser } from "./actions";

//auth options to be used by the next-auth authentication
export const authOptions: NextAuthOptions = {
  providers: [
    //list of providers for authentication
    GoogleProvider({
      clientId: process.env.GOOGLE_CLIENT_ID!,
      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
    }),
  ],
  jwt: {
    //jwt encoding and decoding properties of auth options
    encode: ({ secret, token }) => {
      const encodedToken = jsonwebtoken.sign(
        {
          ...token,
          iss: "grafbase",
          exp: Math.floor(Date.now() / 1000) + 60 * 60,
        },
        secret
      );

      return encodedToken;
    },
    decode: async ({ secret, token }) => {
      const decodedToken = jsonwebtoken.verify(token!, secret);
      return decodedToken as JWT;
    },
  },
  theme: {
    colorScheme: "light",
    logo: "/logo.svg",
  },
  callbacks: {
    async session({ session }) {
      //callback function called everytime the user visits the web page and get user data
      const email = session?.user?.email as string;
      try {
        const data = (await getUser(email)) as { user?: UserProfile };

        //create new user session with data from database
        const newSession = {
          ...session,//return the already available session from google
          user: {//add user data to the new session
            ...session.user,
            ...data?.user,
          },
        };

        return newSession;
      } catch (error: any) {
        console.error("Error retrieving user data: ", error.message);
        return session;
      }
      return session;
    },
    async signIn({ user }: { user: AdapterUser | User }) {
      //callback function called wen user signs in google user or user in database
      try {
        //1st try to get the user from database if user exist
        const userExists = (await getUser(user?.email as string)) as {
          user?: UserProfile;
        };

        //if not we create the user
        if (!userExists.user) {
          await createUser(
            user.name as string,
            user.email as string,
            user.image as string
          );
        }

        return true;

        return true; //if user is created
      } catch (error: any) {
        console.log("Error checking if user exists: ", error.message);
        return false;
      }
    },
  },
};

export async function getCurrentUser() {
  const session = (await getServerSession(authOptions)) as SessionInterface;

  return session;
}
